+++ projects/iad-interface-admin/form/src/lib/projection-form/iad-reference-projection-provider.interface.ts
@@ -2,5 +2,10 @@ import { Observable } from 'rxjs';
 import { IadFormProjectionInterface } from './model/iad-form-projection.model';
 
 export interface IadReferenceProjectionProviderInterface {
-    findProjectionsByName(input: { [param: string]: string | string[] }): Observable<{ [param: string]: IadFormProjectionInterface }>;
+
+    /**
+     * Will allow to get form projection map from API
+     * @param input
+     */
+    findFormProjectionsByName?: (input: { [param: string]: string | string[] }) => Observable<{ [param: string]: IadFormProjectionInterface }>;
 }