+++ projects/iad-interface-admin/form/src/lib/projection-form/public-services/iad-reference-projection-provider.service.ts
@@ -7,7 +7,7 @@ import { IadReferenceProjectionProviderInterface } from '../iad-reference-projec
     providedIn: 'root'
 })
 export class IadReferenceProjectionProviderService implements IadReferenceProjectionProviderInterface {
-    findProjectionsByName(input: { [p: string]: string | string[] }): Observable<{ [p: string]: IadFormProjectionInterface }> {
+    findFormProjectionsByName(input: { [p: string]: string | string[] }): Observable<{ [p: string]: IadFormProjectionInterface }> {
         console.error('IadReferenceProjectionProviderService is not implemented!');
         return undefined;
     }