+++ projects/iad-interface-admin/src/public_api.ts
@@ -5,23 +5,16 @@
 export * from './lib/model/iad-interfaces';
 export * from './lib/model/iad-model';
 
-// #Service
-export * from './lib/public-services/iad-projection-load.service';
-
 // #Utils
 
 // #Common
-export * from './lib/common/tabbed-view/projection-tab.model';
-export * from './lib/common/tabbed-view/tabbed-view.component';
 export * from './lib/common/iad-common.module';
-
 export * from './lib/common/helpers/projections.helper';
 
 // # base grid:
 export * from './lib/iad-base-grid/model/iad-grid-column.model';
 export * from './lib/iad-base-grid/model/iad-grid-model';
 export * from './lib/iad-base-grid/model/column-components.model';
-export * from './lib/iad-base-grid/base-grid/cms-setting';
 export * from './lib/iad-base-grid/base-grid/base-grid-columns.service';
 export * from './lib/iad-base-grid/table-header-menu/table-header-menu.component';
 export * from './lib/iad-base-grid/base-grid/base-grid.component';